Challenge 스테이지 클리어 시 패킷 전송 구현, 패킷 수정,

+++ Client/Capstone Project/Assets/Scripts/Challenge_Progress.cs
@@ -9,8 +9,11 @@ public class Challenge_Progress : MonoBehaviour
     Transform _challenge = GameObject.Find("ChallengeStage_Select").transform;
     void Start()
     {
-        ushort a = 0;
+
+        byte a = 0;
+
         for (ushort i = 1; i <= 10; i++)
+
         {
             if (Managers.User.ChallangeStageInfo.TryGetValue(i, out a))
             {
